Revert "Use JWTAuth Middleware to check correct authentication with the backend. (#872)"
This reverts commit 1a25394c1d65c8828559c75570ccf368994f7926.
