+++ fastlane_app.rb
@@ -12,6 +12,7 @@ require_relative "app/shared/logging_module"
 require_relative "app/shared/dot_keys_variables"
 require_relative "app/shared/fastlane_ci_error" # TODO: move somewhere else
 require_relative "app/features/build_runner/build_runner"
+require_relative "app/features-json/jwt_auth"
 
 # All things fastlane ci related go in this module
 module FastlaneCI
